add symbol visibility stuff

+++ test/dtor/ref/dtor-sys/src/std_string.rs
@@ -42,7 +42,7 @@ pub fn std____cxx11__basic_string_char__copy(this_: *mut std___cxx11_string_t, r
 
 pub fn std____cxx11__basic_string_char__dtor(this_: *mut std___cxx11_string_t) -> Exception;
 
-pub fn std____cxx11__basic_string_char__assign(this_: *mut std___cxx11_string_t, return_: *mut *mut std___cxx11_string_t, s: *const c_char, count: c_ulong) -> Exception;
+pub fn std____cxx11__basic_string_char__assign(this_: *mut std___cxx11_string_t, return_: *mut *mut std___cxx11_string_t, s: *const c_char, count: usize) -> Exception;
 
 pub fn std____cxx11__basic_string_char__c_str(this_: *const std___cxx11_string_t, return_: *mut *const c_char) -> Exception;
