+++ test/std/ref/std-sys/src/std_set.rs
@@ -54,7 +54,7 @@ pub fn std__set_std__string__cbegin(this_: *const std_set_string_t, return_: *mu
 
 pub fn std__set_std__string__cend(this_: *const std_set_string_t, return_: *mut std_set_string_iterator_t) -> Exception;
 
-pub fn std__set_std__string__size(this_: *const std_set_string_t, return_: *mut c_ulong) -> Exception;
+pub fn std__set_std__string__size(this_: *const std_set_string_t, return_: *mut usize) -> Exception;
 
 pub fn std___Rb_tree_const_iterator_std____cxx11__basic_string_char___deref(this_: *const std_set_string_iterator_t, return_: *mut *const std_string_t) -> Exception;
