+++ test/std/ref/std-sys/src/std_string.rs
@@ -59,7 +59,7 @@ pub fn std____cxx11__basic_string_char__sizeof() -> usize;
 
 pub fn std____cxx11__basic_string_char__alignof() -> usize;
 
-pub fn std____cxx11__basic_string_char__assign(this_: *mut std_string_t, return_: *mut *mut std_string_t, s: *const c_char, count: c_ulong) -> Exception;
+pub fn std____cxx11__basic_string_char__assign(this_: *mut std_string_t, return_: *mut *mut std_string_t, s: *const c_char, count: usize) -> Exception;
 
 pub fn std____cxx11__basic_string_char__c_str(this_: *const std_string_t, return_: *mut *const c_char) -> Exception;
